Use 'git checkout' when checking out a tag
The way we were checking out tags did not match the results that would
happen from a simple 'git checkout tagname', which made it difficult for
us to push repos from the executor to the workers.  It may also have
eventually caused inconsistent results with git tooling in jobs since
it didn't match the way most folks would check out a tag.  (In particular,
this method did not appear to update .git/logs/HEAD, which is used by
git status.)

Change to just use the git checkout command rather than the lower-level
manipulation we were doing.

Change-Id: I77e9900c73caf16032dbadb9991358caed0e07ed