Emit shell instead of script tasks
Unmatched single quotes is what killed us before. Add an option which
does a test with ansible-playbook --syntax-check to validate whether a
playbook emitted with shell instead of script works.

Running that way takes around 90 minutes. So default to shell, but
provide --syntax-check as a way to check using shell and fallback to
script. This can then be used to fix pre-migration jobs if desired.

Also separate task entries with an extra '\n' for readability.

And only collect the generated playbooks - not all of them.

Change-Id: I147a7562e6578ab300ef56ae7b1c9d9d3f8f920f