Fix infinite loop on reconfiguration exception
If there was an error re-enqueing a change during reconfiguration,
we would abort reconfiguration after having moved some changes into
the new pipelines.  Once we start doing this, we need to finish it,
otherwise items will be in change queues which are no longer attached
to the pipelines in use.

To avoid the issue, simply remove an item from the pipeline if an
exception is raised while re-enqueing it.  This is bad, but not as
bad as an infinite loop.

Change-Id: I628e599671b7d2d14ae981d65d04e37b23d26831