Fix race in test_in_repo_branch
We've told zuul about a new branch, so it's going to go look for
configuration in it, but while that's happening, we're adding a
new change upstream.  Adding the change will alter the upstream
repo sufficiently that queries to the fake gerrit to ask about
refs are likely to be affected by the process of adding the change.

To correct, wait for the system to stabilize after adding the branch
and before adding the new change.

Change-Id: Ie3e2e9b5ad35ba99c32261ab8e9b6265a95b7cde