Run ansible setup at start of job to verify network
Run the Ansible 'setup' module on all hosts in the inventory
at the start of the job with a 60 second timeout.  If we
aren't able to connect to all the hosts and gather facts
within that timeout, there is likely a network problem
between here and the hosts in the inventory; return them and
reschedule the job.

Change-Id: I121514aa1f3c9ac1a664322c8d7624703146f52f