Always try to unlock nodes when returning
Even if things are not in an expected state (eg, node not locked)
or we encounter errors updating node state, always try to unlock
the nodes when returning a nodeset.  Finally, if that too fails,
it will be logged but not treated as a fatal exception so that
we don't get stuck in queue processing.

Change-Id: Iab3bae4146c62067342e0c521a6289d045e65a5c