Fix Gearman UnknownJob handler
In the executor client when we find an unknown job we attempt to iterate
over the known builds to match up the job handles (to identify the
unknown job). Unfortuantely we iterate over the keys of the builds dict
which is a string uuid rather than a build object. Iterate over the
values() which are build objects instead.

Change-Id: I865eff0c48d9e986fc040a334cad85d393cbff0b