Move shadow layout to item
It turns out layouts can be large[citation needed].  The layout used
for a prior buildset is only of historic interest.  The shadow layout
is effectively used only to freeze a job graph, and once a buildset
has been superceded, that is no longer necessary.  Drop the reference
at this point to allow the memory to be reclaimed sooner.

Change-Id: I2bf651ab99f1b1bfbe0e2442cf26c40475022fe5