Don't modify project-templates to add job name attributes
A recent change began performing full validation on project-pipeline
job variants rather than deferring to the jobparser.  However,
we were modifying the config object inside the project-template parser
to add the job name, before passing it to the job parser.  This meant
that the template passed validation the first time, but during a
subsequent reconfiguration, the cached value would already have the
name attribute and fail validation.

This change stops adding the name attribute, instead passing it to
the jobparser.  It also disables validation by the jobparser in the
case that the project-template parser has already performed it, so that
we don't slow reconfiguration.

Change-Id: I080b4dd01ac40363a932c0853f0f9670b1e2c511