Don't use cached config when deleting files
During dynamic reconfiguration, we treated deleted files the same
as files which were not present in the collected file set for the
changes in question.  This meant that if someone deleted a file
from a project-branch, we would end up using the cached data rather
than no data.

Only use cached data if there is no entry for the project-branch
in the files object.  If there is an entry, even if it's empty,
do not use the cached data.

Change-Id: If18bfa12b6c8e9ac5733bfe597f48b90ec49df9e