Mask dependent config errors
When a change to an untrusted project depends on a change to a
config project, we can validate that the resulting config will be
correct, but we can't use it.  Currently, in that case, we report
the configuration "error" that results from not using the
dependent config-project change.  Instead, leave a message for the
user saying that everything will be okay if they wait for that
change to merge.

Change-Id: I34a002000d1bf640b298b7bf72d1af5a3a72ced1