Make statsd import failure raise exception
At this point, if a user has configured statsd, we assume we can
import it, so don't wrap the import in a try_import, just assume
it's going to work.

However, we still import within the method so that if statsd is not
configured, we don't attempt to import it.

Change-Id: I5f0041300c666a3d572f45f3709d5d64ad599c8a