Ignore parents on job variants
Job variants generally should not set the parent attribute.  Instead,
the inheritance path of the reference definition of the job should be
used.

I had previously considered honoring parent in variants a harmless
undocumented feature (which permitted a sort of crude
multiple-inheritance).  However, this makes it very difficult, if not
impossible, to create correct branch variants of jobs from the initial
branch point -- a new branch will end up with an identical copy of the
jobs from the master branch, including the parent attribute.  Since
the behavior difference between the jobs on the two branches runs counter
to what the user would expect, let's ignore the parent attribute on
variants.

Change-Id: I85ffb014c73e39631c76debb04d8c9775ab097ad