Disable sql_reporter alembic downgrades
When I built the sql_reporter I had never intended to support alembic
migration downgrades. These are problematic for a number of reasons
(that other projects such as nova have discovered in the past).

The main problem with downgrades is that if you have a db version
checked out, you also have the code that requires this version. That
is, you should not downgrade removing a column or table that the code
expects.

Instead of downgrades operators should be encrouaged to take backups
and restore from a previous point (along with rolling back the code)
if they wish to undo an upgrade.

Change-Id: I8417de803040dd11db2875e2069f80feb05df19f