Don't set the build result to ABORTED
We pass 'ABORTED' back from the executor in some cases, mostly to
aid in testing (so we know when the executor has intentionally
aborted a job).  However, that's a non-final result, and other
parts of zuul (like areAllJobsCompleted) shouldn't need to know
about it.  If the executor client receives ABORTED (or DISCONNECT),
log it, but don't pass it any further up the chain.

Change-Id: If2cbf2f6eecb9a7fdef4d0a6eebeb5fe2549e276