Default change and patchset to NULL in SQLReporter
The schema for the zuul_buildset table has change and patchset
columns as an integer type, so use a NULL default for them when
reporting builds which lack actual data for these fields (e.g.,
those triggered by Gerrit ref-updated events). Previously this
was attempted with an empty string, but the ORM does not coerce
that to NULL so we must use None instead to achieve that.

Change-Id: Ifaa22035182f0848248f394a8d9d4b278ff23583