Correctly stream the remaining buffer
In case ansible crashes while executing a command module we loose the
last line of the log. This is caused because we don't stream the
remaining buffer but the last line we split out of the buffer.

I wasn't able to construct a test case which makes the original code
break thus this change is without a test case.

Change-Id: I10ffe248c756e0a77f39f6e075dd3615c53f1b03